Fixed issue introduced in 50470d30513376e28b8d604f19845905616b9201 The game was freezing when the game prints "Take good care of [whatever]" as the trade finished. Thanks to MGriffin, we found the reason why; the ribbon changes made in the commit linked in the title. There's a function that seems to check if a Pokémon that is received during a trade has a special ribbon. A ribbon for being in the party that defeated the Pokémon League, a ribbon awarded for completing 50 smth battles in the Battle Tower and such. This function has a hardcoded check, and that was the root of the problem.

+++ src/pokemon_summary_screen.c
@@ -166,9 +166,9 @@ static EWRAM_DATA struct PokemonSummaryScreenData
         u8 sanity; // 0x35
         u8 OTName[17]; // 0x36
         u32 OTID; // 0x48
-        u8 dayMet;
-        u8 monthMet;
-        u8 yearMet;
+        u32 dayMet; // 0x4C
+        u32 monthMet; // 0x50
+        u32 yearMet; // 0x54
     } summary;
     u16 bgTilemapBuffers[PSS_PAGE_COUNT][2][0x400];
     u8 mode;
